# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SaveUpdateResProduct": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"ProductName": {
						"type": "string"
					},
					"ProductType": {
						"type": "string"
					},
					"ProductPrice": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"ProductName",
					"ProductType",
					"ProductPrice",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveProduct": {
				"properties": {
					"ProductName": {
						"type": "string"
					},
					"ProductType": {
						"type": "string"
					},
					"ProductPrice": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ProductName",
					"ProductType",
					"ProductPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetProduct": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProduct": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"ProductName": {
						"type": "string"
					},
					"ProductType": {
						"type": "string"
					},
					"ProductPrice": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"ProductName",
					"ProductType",
					"ProductPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteProduct": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResOrder": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"User_email": {
						"additionalProperties": false,
						"type": "object"
					},
					"Product": {
						"anyOf": [
							{
								"type": "string"
							},
							{}
						]
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"User_email",
					"Product",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SingleOrderProduct": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Product_quantity": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"Product_quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveOrder": {
				"properties": {
					"Product": {
						"items": {
							"$ref": "#/components/schemas/SingleOrderProduct"
						},
						"type": "array"
					},
					"Deleivery_address": {
						"type": "string"
					},
					"User_contact_number": {
						"type": "string"
					}
				},
				"required": [
					"Product",
					"Deleivery_address",
					"User_contact_number"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetOrder": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateOrder": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Product": {
						"items": {
							"$ref": "#/components/schemas/SingleOrderProduct"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"Product"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteOrder": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResUser": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"email",
					"password",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqUser": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isAdmin": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"isAdmin"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetReqUser": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"getReqAuth": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "http",
				"name": "Authorization",
				"in": "header",
				"scheme": "Bearer"
			}
		}
	},
	"info": {
		"title": "webapi",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/Product/savenewproduct": {
			"post": {
				"operationId": "Saveproduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResProduct"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveProduct"
							}
						}
					}
				}
			}
		},
		"/Product/getproduct": {
			"post": {
				"operationId": "Getproduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResProduct"
								}
							}
						}
					}
				},
				"summary": "Only admin can see particular product",
				"tags": [
					"Product"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetProduct"
							}
						}
					}
				}
			}
		},
		"/Product/updateproduct": {
			"put": {
				"operationId": "Updateproduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResProduct"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProduct"
							}
						}
					}
				}
			}
		},
		"/Product/deleteproduct": {
			"delete": {
				"operationId": "Deleteproduct",
				"responses": {
					"200": {
						"description": "Product deleted",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteProduct"
							}
						}
					}
				}
			}
		},
		"/Product/getproductslist": {
			"post": {
				"operationId": "Getproductslist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResProduct"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/Order/saveneworder": {
			"post": {
				"operationId": "Saveorder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResOrder"
								}
							}
						}
					}
				},
				"summary": "You have to login first so your id will also be saved",
				"tags": [
					"Order"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveOrder"
							}
						}
					}
				}
			}
		},
		"/Order/getorder": {
			"post": {
				"operationId": "Getorder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResOrder"
								}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetOrder"
							}
						}
					}
				}
			}
		},
		"/Order/updateorder": {
			"put": {
				"operationId": "Updateorder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResOrder"
								}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateOrder"
							}
						}
					}
				}
			}
		},
		"/Order/deleteorder": {
			"delete": {
				"operationId": "Deleteorder",
				"responses": {
					"200": {
						"description": "Order deleted",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteOrder"
							}
						}
					}
				}
			}
		},
		"/Order/getorderslist": {
			"post": {
				"operationId": "Getorderslist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "User can only see his orders",
				"tags": [
					"Order"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/Order/getorderbill": {
			"post": {
				"operationId": "Getorderbill",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResOrder"
								}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetOrder"
							}
						}
					}
				}
			}
		},
		"/User/savenewuser": {
			"post": {
				"operationId": "Saveuser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResUser"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqUser"
							}
						}
					}
				}
			}
		},
		"/User/getuser": {
			"post": {
				"operationId": "Getuser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResUser"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetReqUser"
							}
						}
					}
				}
			}
		},
		"/Login/getauth": {
			"post": {
				"operationId": "Getauthuser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/getReqAuth"
								}
							}
						}
					}
				},
				"tags": [
					"Login"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}